move conversion-base to simplify pr diff
