+++ crypto/bindings/conversion-proof.ts
@@ -26,7 +26,7 @@ import {
   fieldFromRust,
   fieldsToRustFlat,
   fieldsFromRustFlat,
-} from './conversion-base.js';
+} from './conversion-base-old.js';
 import { ConversionCore, ConversionCores } from './conversion-core.js';
 
 export { proofConversion };
